Fixed type of waterfall image

+++ detection/change_detection/scripts/change_detector.py
@@ -127,7 +127,7 @@ class ChangeDetector(object):
         for i in range(np.size(im_with_keypoints,2)):
             f = scipy.interpolate.RectBivariateSpline(np.linspace(0 ,255, np.size(im_with_keypoints, 0)), np.linspace(0,255, np.size(im_with_keypoints, 1)), im_with_keypoints[:,:,i])
             out_img_array[:,:,i] =  f(np.linspace(0, 255, np.size(img_array, 0)), np.linspace(0, 255, np.size(img_array, 1)))
-        out_img_array = out_img_array.astype(float)
+        out_img_array = out_img_array.astype(np.uint8)
         return out_img_array
 
     def pcloud2ranges(self, point_cloud, tf_mat):
